feat: add separate .js files for workSuggestion and leisureSuggestion
