+++ src/App.js
@@ -1,82 +1,18 @@
 import React, { useState } from "react";
 import "./App.css";
 import MovieSuggestion from "./MovieSuggestion";
+import leisureSuggestion from "./leisureSuggestion";
+import workSuggestion from "./workSuggestion";
 
 
 // Leisure activity suggestions based on time + energy
-const leisureSuggestions = {
-  mini_low: [
-    "Do a 4-7-8 breathing cycle",
-    "Smell your favorite candle or perfume",
-    "Look at funny memes or photos",
-    "Close your eyes and listen to your surroundings for 1 minute",
-    "Do 10 seconds of deep stretching"
-  ],
-  short_medium: [
-    "Sketch a doodle with no pressure",
-    "Try a 10-minute guided meditation",
-    "Water your plants",
-    "Make a comfort playlist",
-    "Watch a YouTube video"
-  ],
-  medium_high: [
-    "Put on music and dance it out",
-    "Take a walk around the block",
-    "Cook something fun but simple",
-    "Rearrange a small corner of your space",
-    "Call a loved one"
-  ],
-  long_any: [
-    "Read a book or comic you enjoy",
-    "Try a new recipe",
-    "Watch a movie",
-    "Do a slow creative task like painting",
-    "Play a cozy video game"
-  ]
-};
+    return (
+  <LeisureSuggestion />
+    );
 // work activity suggestions based on mood + energy
-const workSuggestions = {
-  calm_low: [
-    "Journal quietly for 5 minutes",
-    "Organize your thoughts in a notes app",
-    "Reflect on a recent win or lesson"
-  ],
-  calm_high: [
-    "Start a creative project",
-    "Outline a blog post or brainstorm ideas",
-    "Work on a long-term goal"
-  ],
-  anxious_low: [
-    "Do a 5-minute brain dump",
-    "Give your workspace a light clean",
-    "Take deep breaths and jot down your work worries"
-  ],
-  anxious_high: [
-    "Tackle a simple, low-stakes task",
-    "Organize your email inbox",
-    "Sort your digital files or desktop"
-  ],
-  bored_medium: [
-    "Try a productivity technique like Pomodoro",
-    "Set a silly challenge (e.g. write 10 ideas in 5 minutes)",
-    "Reorganize your workspace layout"
-  ],
-  restless_high: [
-    "Do a standing task (like writing ideas on a whiteboard)",
-    "Move while working (pace, stretch, stand)",
-    "Do quick research for a fun side project"
-  ],
-  motivated_high: [
-    "Start your most important task right away",
-    "Work on something with a deadline",
-    "Make progress on a passion project"
-  ],
-  default: [
-    "Make a short to-do list",
-    "Check your calendar or planner",
-    "Spend 5 minutes making a plan"
-  ]
-};
+    return (
+  <WorkSuggestion />
+    );
